// src/api/model/clientDto.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface ClientDto { 
    adress?: string;
    cpf?: string;
    fone?: string;
    id?: number;
    name?: string;
    rg?: string;
}